+++ doc/mysql-schema/frostmourne.sql
@@ -3,7 +3,7 @@ CREATE DATABASE frostmourne
     DEFAULT COLLATE utf8mb4_0900_ai_ci;
 
 /* if your mysql not support utf8mb4_0900_ai_ci, use collate utf8mb4_general_ci instead */
---CREATE DATABASE frostmourne DEFAULT CHARACTER SET utf8mb4 DEFAULT COLLATE utf8mb4_general_ci;
+-- CREATE DATABASE frostmourne DEFAULT CHARACTER SET utf8mb4 DEFAULT COLLATE utf8mb4_general_ci;